apps/agent/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
langchain>=0.1.13
langchain-aws>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.2.2
python-dotenv>=1.0.1
pydantic>=2.6.4
boto3>=1.34.69
